lisa/requirements/heroku.txt
-r common.txt

dj-database-url==0.3.0
dj-static==0.0.6
static3==0.6.1
whitenoise==3.3.1
gunicorn==19.7.1
python-dotenv==0.10.3
psycopg2-binary==2.8.3
